Encode + for oauth-1.0a
This commit fixes #90 by making sure that spaces in URLs are encoded
using %20 instead +. While technically correct the underlying oauth-1.0a
library will encode a + as %2B when signing, which causes an 401
Authentication error, because the URL that was signed was not the URL
that was fetched.

The fix isn't particulary satisfying, but it does pass the tests except
for `should stream data without any rules` which was already failing.